Refactored s:ConstructAndExecuteCmd to not require reload param.
Previously the reload parameter was needed as the
git command along with the --git-dir flag was
stored and re-executed after being modified for a
new commit count. Due to naming the buffer
the command can now work out the git-dir each time and
thus this isn't needed. This should simplify the
logic here a lot.