Use Ctrl-n/p to navigate revisions
Map C-n/C-p to navigate down/up through revisions. Moves down/up through
the list of revisions (in file and browser mode) and loads that
revision. Can be very useful to see changes in a file. (Unfortunately,
it doesn't keep you at the same point in the file, which would we even
better.)