Prevent bdelete errors if buffer doesn't exist
We don't need bdelete to complain if the fugitive buffer has already been
deleted. We're doing this as magic in the background to make sure that
the buffer list isn't polluted with all the fugitive windows we open, so
it's okay if the buffer was already deleted.

This error is prevalent if the user sets fugitive buffers to auto-delete
on close:
    autocmd BufReadPost fugitive://* set bufhidden=delete