Open commit instead of file to create a tab
It appears that the only reason we use Gtabedit is to get a new tab
where fugitive functions will work correctly. HEAD is a real file, so
editing it doesn't work, but the tree referenced by HEAD is not, so
we have a stronger guarantee that it will work correctly.

When we edit HEAD, we are editing .../.git/HEAD, but HEAD:
doesn't actually exist so fugitive can't be confused by the path.

Related to tpope/vim-fugitive/issues/309
Fixes #40.