Fix error in JumpToCommit.
This fixes the "Invalid Argument" error generated if you try to use C-p
to go back up a commit in Browser view as of e9b71cca44.
I'm not sure if this was tested against an old version, but at least in
my version of vim (7.3) the operator to concatenate strings is '.=', not
'+='.