Add command to jump to parent commit
Map the "p" key to make it easy to jump to a parent of the current
commit, even if there are many intervening commits on other branches,
loading enough of the history to include the target commit.  Allow a
count to be used to specify a parent other than the first for merge
commits.